stagging/Full_NAME/requirements.txt
flask==3.0.2
python-dotenv==1.0.1
google-generativeai==0.8.4
requests==2.31.0
werkzeug==3.0.1
jinja2==3.1.3
click==8.1.7
itsdangerous==2.1.2
certifi==2025.1.31
urllib3==2.2.1
charset-normalizer==3.4.1
idna==3.10
jsonschema==4.21.1
tqdm==4.66.2 